requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
chromadb>=0.5.0
pypdf>=4.2.0
python-docx>=1.1.0
tiktoken>=0.7.0
openai>=1.35.0
python-dotenv>=1.0.1
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0